move query_sql to main linker object
