+++ splink/duckdb/duckdb_linker.py
@@ -18,8 +18,6 @@ from ..logging_messages import execute_sql_logging_message_info, log_sql
 from ..misc import (
     ensure_is_list,
     all_letter_combos,
-    _check_dependency_installed,
-    query_sql_to_splink_df,
 )
 from ..input_column import InputColumn
 
@@ -47,23 +45,20 @@ class DuckDBLinkerDataFrame(SplinkDataFrame):
 
         self.duckdb_linker._delete_table_from_database(self.physical_name)
 
-    def as_dataframe(self, output_type, limit=None):
+    def as_record_dict(self, limit=None):
+
         sql = f"select * from {self.physical_name}"
         if limit:
             sql += f" limit {limit}"
 
-        return self.duckdb_linker.query_sql(sql, output_type)
-
-    def as_record_dict(self, limit=None):
-        return self.as_dataframe(output_type="pandas", limit=limit).to_dict(
-            orient="records"
-        )
+        return self.duckdb_linker._con.query(sql).to_df().to_dict(orient="records")
 
     def as_pandas_dataframe(self, limit=None):
-        return self.as_dataframe(output_type="pandas", limit=limit)
+        sql = f"select * from {self.physical_name}"
+        if limit:
+            sql += f" limit {limit}"
 
-    def as_arrow_table(self, limit=None):
-        return self.as_dataframe(output_type="arrow", limit=limit)
+        return self.duckdb_linker._con.query(sql).to_df()
 
 
 class DuckDBLinker(Linker):
@@ -196,20 +191,6 @@ class DuckDBLinker(Linker):
         self._con.execute(sql).fetch_df()
 
         return DuckDBLinkerDataFrame(templated_name, physical_name, self)
-
-    def query_sql(self, sql, output_type="pandas"):
-        if output_type in ("splink_df", "splinkdf"):
-            return query_sql_to_splink_df(self, sql)
-        elif output_type == "pandas":
-            return self._con.execute(sql).fetch_df()
-        elif output_type in ("arrow", "pyarrow"):
-            _check_dependency_installed("pyarrow")
-            return self._con.execute(sql).arrow()
-        else:
-            raise ValueError(
-                f"output_type '{output_type}' is not supported.",
-                "Must be one of 'splink_df'/'splinkdf', 'pandas' or 'arrow'/'pyarrow'",
-            )
 
     def register_table(self, input, table_name, overwrite=False):
